Add compatible string to psu_ddr and psv_ddr
This patch adds the support to generate compatible string for psu_ddr
and psv_ddr memory nodes. we have already an API gen_compatible_property
to add compatible string, but this takes argument as drv_handle. but the
requirement here is to add compatible property to memory node(memory@0),
which is not drv_handle.

Signed-off-by: Naga Sureshkumar Relli <naga.sureshkumar.relli@xilinx.com>