Removing hardcoded video pipelines for multimedia IPs
From the hdf/xsa there is no sufficient data to create video pipelines
nodes, so from DTG we created the video pipeline nodes by hardcoding the
endpoints and remote endpoints. This is breaking when someone designs
has different combinations of video pipelines rather than the supported
ones in the DTG. Till we get the sufficient endpoint/remote endpoint
information in the hdf/xsa, we are removing the generation of video
pipelines.

Signed-off-by: Venkatesh Yadav Abbarapu <venkatesh.abbarapu@xilinx.com>