Fix v_procss hier IP related issues
The v_procss is hierIP and contains some IPs inside it.
Due to failure in get_mem_ranges command, we are not properly generating
the reg values. so read the BASE_ADDR if get_mem_ranges is failed and the
generate the reg value.

Also there is a base address mismatch for PS UART, psu_iou_s and zynq_ultra_ps_e.
for all these three instances the base address is same which is 0xFF000000.
because of this the reg is generating multiple times. so for time being we are
blocking thepsu_iou_s and zynq_ultra_ps_e to not to generate the nodes for these.

Signed-off-by: Naga Sureshkumar Relli <naga.sureshkumar.relli@xilinx.com>