Add cascade mode support for axi intc controllers
The cascade mode can be used when the processor has more than 32 interrupts.
A single instance of axi intc can handle maximum of 32 interrupts. In
cascade mode, there are two or more axi intc instances connected to
a processor. The properties C_EN_CASCADE_MODE and C_CASCADE_MASTER
are used. When the C_EN_CASCADE_MODE=1 and C_CASCADE_MASTER=1
is for primary instance of the axi intc, bit 31 of intr is the
cascaded interrupt port. When the C_EN_CASCADE_MODE=1 and C_CASCADE_MASTER=0
the intr ports are connected to primary instance of axi intc core, bit 31 of
intr is cascaded interrupt port. When the C_EN_CASCADE_MODE=0 and C_CASCADE_MASTER=0
means no cascade interrupt feature available.

Signed-off-by: Venkatesh Yadav Abbarapu <venkatesh.abbarapu@xilinx.com>
Signed-off-by: Naga Sureshkumar Relli <naga.sureshkumar.relli@xilinx.com>