Remove duplicate node addresses
In some designs same IP is mapped with same address is mapped
to different processors. in this case the node is getting generated
based on the number of instances that is mapped to the processor.
in this case we are getting dtc error "duplicate node names".
So this patch fixes this by not generating the same node again
even though the instane name is different.

Signed-off-by: Naga Sureshkumar Relli <naga.sureshkumar.relli@xilinx.com>