Add checks to remove duplicate nodes
In some designs, one IP can be mapped to different masters.
Ex: a mailbox can be mapped to different masters.
In this case the IP is same but the instance name is different
for each master. so with this we see dtc errors "duplicate nodes".
So fix by just generating the one instance, because the base
address is anyway same.

Signed-off-by: Naga Sureshkumar Relli <naga.sureshkumar.relli@xilinx.com>