Add status okay to ams_pl and ams_ps
The ams_ps and ams_pl are child nodes to ams node and are not part
of the XSA. Hence we are not generating status okay to ams_pl and ams_ps.
Fix this to generate status okay in pcw.dtsi through ams driver tcl.
This patch also adds the ams_ps and ams_pl entries to cpu cluster
address-map, because the ams_ps and ams_pl are not part of XSA, hence
address-map doesn't contain these ams_ps and ams_pl.

Signed-off-by: Naga Sureshkumar Relli <naga.sureshkumar.relli@xilinx.com>