common_proc: Add fifo_generator to the list of ignore_list ip's to avoid duplicate names
When there are mutiple instances of same ip, the node name for each instance
should be generated with unqiue name and it was happening for most of the
ip's with <name>@<address> format but there are few ip's which dont have
address associated with it and in those cases those are getting generated
with name <name>@<count> where the count will be incremented for each instance.

There is already a list of such ip's maintained in common_proc tcl and this
patch adds fifo_generator also to that list inorder to avoid duplicate node
names being generated when there are multiple insances of fifo_generator.

Signed-off-by: Siva Durga Prasad Paladugu <siva.durga.paladugu@xilinx.com>