common_proc: Fix type for driver properties
For some properties hex decimal string can come in between a string
and in those scenarios, the existing code to define thetype assumes it to be
hexint type but that doesnt hold true always as the hex string may contain in
between a string in which case the type for that property should be string instead
of hexint. This patch fixes this issue by ensuring that hexa decimal should be at
start of the string for it to be defined as hexint type otherwise it will be treated
as string type.

Signed-off-by: Siva Durga Prasad Paladugu <siva.durga.paladugu@xilinx.com>