system-device-tree-xlnx: device_tree: Skip device tree entry for gmii2rgmii
System-device-tree-xlnx generates device tree nodes for all the IPs
present in the design. When gmii2rgmii node is present in the design,
along with emacps, emacps tcl adds a device tree entry for gmii2rgmii,
under its mdio node. SDT also generates another node for the same IP,
in pl.dtsi. To avoid multiple nodes corresponding to the same instance
of the IP, skip generating the node in pl.dtsi.

Signed-off-by: Pranavi Somisetty <pranavi.somisetty@amd.com>