device_tree : Avoid adding 40 bit addresses to the 32 bit processors
R5/PMU/PMC/PSM address maps are of 32 bits. When a higher DDR region (with
40 bit address) is mapped to these processors, the addresses are getting
manipulated wrongly, resulting into incorrect NOC_DDR addresses in these
processor's address maps. A new API is therefore added to the common_proc
to check if the address is a 40 bit address and the address is added to
these processor's address map only when the API returns False.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>