ddrpsv: Avoid adding the DDR addresses to the processor address map if no region is mapped to the processor
There is a try catch block when the processor mapping for a DDR IP is being checked.
When try statement fails, the catch statement is reporting wrong base_address and
high_address of the DDR IP. This is fixed by getting the memory peripheral mapped
to the processor in the beginning itself. If it returns nothing, the rest of the
steps is avoided for that processor. Furthermore, redundant usage of variable to
get the peripheral name is also removed (by removing slave variable and moving
periph variable at the top).

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>