ddrpsv: Update in logic to get the base_address and high_address of each DDR region
There is a difference in the logic of DTG and the embeddedsw repo ddrpsv tcl
that is being used to get base_address and high_address of each DDR region.
If there are multuiple DDR regions with same name (and different master interface),
then the union of all the DDR region with the same name (regardless of master
interface) should be accounted for to determine the base address and the high
address of that particular DDR region. This logic is being used in the embeddedsw
ddrpsv tcl. But in existing implementation (and in DTG), the base address is
never updated once that DDR region is accessed. Therefore, a new proc
"ddrpsv_handle_address_details" is added (similar to what we have in esw) to get the
union of all the address maps for a particular DDR region.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>