common: Adding a catch statement when an env var is set to empty
When a env variable is set to an empty string "", Windows treat this env var as
a undefined variable. On the other hand, linux accepts this assignment with the
empty string. A proper check is needed for such env variables before assigning
it to a local variable.
Reference link: https://wiki.tcl-lang.org/page/env

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>