video_utils : Segragate all the video related procs into a separate file
There are a bunch of procs in common_proc.tcl which are very specific to
the video IP use cases. These procs are very large and repetitive in
nature and tough to maintain. There is a need to separate these application
specific proc into a separate file for better maintenance.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>