slcrps: Adding clkc as a reference instead of a new node inside slcrps
Few properties of clkc are getting generated in pcw.dtsi inside the slcrps node.
Since, clkc node is already present in zynq-7000.dtsi inside slcrps, this new
definition in pcw.dtsi is causing duplicate label error with lopper. Ideally,
create_node should have taken care of this the same way add_or_get_dt_node
is doing in DTG. But, we dont yet have an API in SDT with the intelligence to
find if the label is already generated before. Therefore, to avoid the usage of
create_node in this case, the clkc reference is directly inserted at the end of
pcwdt tree.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>