common_proc: setting interrupt-parent as intc for zynq ps gic
gic label in the legacy linux device tree for zynq as well as DTG zynq-7000.dtsi
is named as "intc". Moreover, the intc_name variable is currently defined only in
zynqmp and versal platform related scopes. Therefore, the interrupt-parent name has
to be changed from ps7_scugic_0 i.e. the IP NAME to intc label and the intc_name
variable needs to be moved out to a common scope.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>