axi_trafgen: convert interrupt missing error from trafgen to a warning
For Linux driver to work, interrupt is mandatory for axi_traffic_generator IP.
Thus, if the interrupt is missing in the design, DTG is throwing an error.
For SDT, the error is not valid and the dt generation has to go through, hence
the existing error is converted into a warning.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>
---
CR: CR-1166306