common_proc/device_tree: Adding logic to decide hw_family for a pure PL design on a SOC
When the design is based out of an SOC say any zynqmp or versal board, family is currently
set to the SOC family. But in case, if the design doesnt have any PS part and is a pure
PL design, the family is still set to the SOC family and there are some hardcoded nodes in
processor cluster (like reset, amba etc) based on these SOC families. Instead family should
be set to microblaze in case of all the pure PL designs irrespective of the SOC family.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>