common_proc/device_tree : Fixing qspi alias issue for SSIT devices
1. Currently addresses for ps mapping is coming from static dtsi files, originating from u-boot
and it is very APU specific. To generate aliases, the code is also looking for those APU mapped
addresses. For pmc_qspi, there can be different MASTER INTERFACEs and hence the BASE_value changes
a/c to those MASTERs. So, just as a workaround for the error during alias mapping for qspi, adding
logic to specifically fetch the A72_0 mapped address of pmc_qspi.

2. Adding the missing slv_psv_pmc entries in the proc_mapping dict.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>