ddrps/ramps: Correcting the ddrps logic and adding ps7_ram node for zynq
ddrps address is presently hardcoded to start with 0 which is not the case in Zynq, it
starts with 0x100000. Similarly, ps7_ram (which is actually an ocm memory in zynq) also
needs to configured according to the design. There are 2 ocm memory comprising of 256 KB
in total. Making changes in tcl to accomodate these features for zynq sdt.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>