common_proc/device_tree : Replacing the get_hw_family with proc driven family decision
The existing logic of get_hw_family() API is relying on FAMILY property of the design.
This is leading to multiple issues while deciding family for PS+PL designs and in adding
designs for new families. Thus, this logic is replaced by Processor driven logic where
the APU processor names will be checked in the design and if none of them are present
and microblaze is present, it will termed as pure microblaze design

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>