common_proc: Adding ps7_qspi_linear to the Memory CNTRL exception list
There are few IPs for which _memory nodes are not explicitly generated inspite of being a MEMORY_CNTRL.
These IPs are being treated like a normal peripheral node as of now (though their identity as a cntrl
and presence in the linker script is something that still needs to be discussed). Adding "ps7_qspi_linear" to
this IP list to map its address to the cpu cluster. Presently it's not getting added as the IP_TYPE is
MEMORY_CNTRL.

Signed-off-by: Appana Durga Kedareswara rao <appana.durga.kedareswara.rao@amd.com>