qspips: Correcting one variable name and the syntax error while adding a comment
There is one condition in qspips which was never tested. Under this condition, spi-max-freq
was being added to the node named as drv_handle which doesnt exist for qspi. Correcting this
typo to use the node corresponding to the drv_handle (i.e. &qspi).

Also, there exists an issue while adding comments in the system device_tree using add_prop API.
Currently, an additional ";" is getting added to the comment starting with /* which is leading to
syntax errors. Adding // in the beginning of the comment as a work around in the interest of time.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>