scuwdt: Fixing the alias issue for scuwdt
scuwdt node is recently added in th zynq-7000 file. As it is a PS IP and already defined
in static PS dtsi file, pcw.dtsi has to use its reference and not create a new node for this.
To check whether the IP is PS or not, gen_ps_mapping API is presently being used, where
the IP's base address is getting matched to a hardcoded dict. Thus, adding the base address
of scuwdt under zynq.dtsi file. Strangely, this duplicate node error is not being reported by
DTC and it is reported only by lopper.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>