ipipsu: Replacing the hardcoded "ipi-target-count" with the design generated values
Number of ipi-targets present within a design is not static and it may vary from design.
As an example, this number comes as 7 for default zynqmp designs and 11 for SSMU CCI enabled
zynqmp designs. Therefore, the hard coded values have been removed from zynqmp and versal
static dtsi files and the logic to generate this value as per the design is being added
in ipipsu specific tcl. While doing so, the stale code in ipipsu.tcl is also being removed.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>
Acked-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>