common_proc: Fix add_cross_property for hex values that are a part of string type property
Inside add_cross_property, there is a condition written to tackle specific use case of 64 bit
addresses (introduced with this commit : de4aaf8d6f3e2). This condition is getting matched even
when the hex values are part of a string and doesnt represent the address (or hex value) as such.
Thus, the condition is modified to check if the value type being added is string or a pure hex
value.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>
Acked-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>