device_tree: Correcting the quote used while checking hierarchical IPs
While generating the address map for hierarchical Video subsystems, a single quote is used to
append those IPs to the overall peripheral list. With this, a pair of  extra quote is getting
appended at the end of the last IP in the periph list when the hsi filter is reporting an empty
string. Converting the single quote to double quote to avoid this issue.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>
Acked-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>