common_proc: Fix the wrongly generated microblaze reg numbers
For all the soft microblaze and pmc/psm/pmu microblazes, the node labels are named
as cpus_microblaze_<num>. The existing logic to get this number is buggy. To fix this,
a new API set_microblaze_list is introduced. Based on the platform, a global list will
be initialized with pmc/psx/pmu microblazes. In addition, all the IP handles with IP_TYPE
as MICROBLAZE will be appended to it. This global list will be used to get the index
for each mb processors wherever needed.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>
Acked-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>