cpu*: Adding a default value of pss-ref-clk-freq property in all the cpu nodes
XPAR_PSU_PSS_REF_CLK_FREQ_HZ macro is needed for most of the boot component applications.
For few designs, these are not coming as part of the IP property. Thus, adding a default
value to the xlnx,pss-ref-clk-freq property of each cpu node. With this, the required
macro will be defined even if the IP doesnt return the needed config value.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>
Acked-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>