device_tree: Ignore coresightps in the PMC/PSM/PMU CPU clusters
PMC/PSM/PMU cpu clusters have coresightps mapped to it as coresight is being treated as
a PS IP and currently all the PS IPs are blindly mapped to the default CPUs if design
is not a Pure PL design. This issue is being fixed by creating a list of IPs that need
to be ignored for a particular CPU cluster and checking these IP's presence before
adding into cluster.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>
Acked-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>