device_tree: Adding check to ignore those IPs that dont have any addresses mapped
There are many IPs which dont have any addresses mapped to it. There is no
mechanism to access drivers for such IPs and generating such nodes is not
adding any value. Therefore, avoid node and property generation of all
such IPs. This will improve timing performance as well as avoid the dtc
compilation errors for duplicate nodes when these IPs are in multiple number.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>
Acked-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>