axi_bram: Correcting the logic to avoid two memory nodes for ilmb and dlmb IPs
We have an API remove_duplicate_addr which is being used to map peripherals
connected at the same address. This is helpful in mapping mailboxes to the
processors as the multiple mailbox can be catering to the same address. The same
API is also mapping the ilmb_ram and dlmb_ram IPs of the same LMB_RAM instance
into one node. While doing so, in case we have two mb instances, each having
bram mapped to it and the LMB_RAM instance of mb_1 having the same base_address 0
as that of mb_0, it combined ilmb_ram and dlmb_ram of both the LMB_RAM instances
into just one node, which altogether led to the missing lmb_ram of microblaze_1.

Therefor, with this patch, lmb_bram_cntrl IP in remove_duplicate_addr API is
ignored and the node mapping logic for ilmb and dlmb_ram is moved to the bram IP
specific axi_bram tcl. In axi_bram tcl, each microblaze memory map is scanned
through, node for any one of the ilmb_ram and dlmb_ram IPs of the same LMB_RAM
instance is only created and the other one is ignored.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>

Acked-for-series: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>