axi_bram: Adding a check if drv_handle from mem map exists
There are few IPs like M_AXI_CTRL_0, APB_M_JESD_RX_0 etc. which are appearing
in the microblaze processor's memory map but are not giving any object output
when hsi get_cells -hier is run over them. This led to an error in the existing
get_property condition. Fixing the same.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>
Acked-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>