ddrps: Replace lsort -integer with lsort -real command
TCL interpreter in Windows OS can't sort more than 32 bit addresses with
lsort -integer command. It is resulting into below error:
"integer value too large to represent"
The same command works fine on Unix systems. One probable reason for this
could be the difference in Word Size. Word Size for tcl_platform is coming
as 8 in Linux and it is 4 for Windows. Thus, replace the lsort -integer
command with lsort -real command which works with both Linux and Windows TCL
interpreter.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>