common_proc.tcl: Stop processing remove_duplicate_addr for hierarchical IPs
In case of Video subsystems, we can have multiple IPs within subsystem having
the same base addresses. And we need all such IPs mapped in the processor
specific address maps. For such cases, we need to check if the IP is located
with a hierarchy and if so, the remove_duplicate_addr API has to avoid
processing them. remove_duplicate_addr API is pecific to mailbox kind of IPs.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>

Acked-for-series: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>