device_tree: Make the gen_afi_node API more modular
There is a lot of repetitive code under gen_afi_node API. Move those part of
code in a for loop to make the code more modular. In addition, create the amba_pl
node in pl.dtsi before proceeding further. This is needed as it is relying on
past APIs to create amba_pl node for it which may fail for few designs. (e.g.
in a generic PS design, amba_pl node is created inside update_endpoints as
axi_noc is the only PL IP in the design and update_endpoints is the only API
that runs for axi_noc).

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>

Acked-for-series: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>