device_tree: Use only the first coresight for serial alias mapping.
Till now, it was assumed that there can only be one coresight in the SOC. Within
P80 devices, now we can have multiple coresights but their usability is still
not clear. Correct the logic to put only the first coresight instance under
alias mapping of the serial for the time being to avoid syntactical issues during
device tree generation.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>

Acked-by: Appana Durga Kedareswara rao <appana.durga.kedareswara.rao@amd.com>