ddrps: Maintain the updated_addr_list as list after updating the high_addr
To get the overall reg property for the memory node (independent of processors),
there is an existing logic to update the detected address list in case when the
current high_addr is less than the incoming high_addr. While updating the list
with lreplace, the list data type is getting messed up and it is coming as two
strings instead of tcl list. Convert the output explicitly to list.

In addition, convert the calculated hex number into caps to have a uniform
representation across the system-top.dts file. Also there was a minor typo under
one of the if condition which will be hit very rarely.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>
Acked-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>