versal.dtsi: Correct the PMC_DMA reg value
The size of PMC_DMA in register space is 0x10000 instead of 0x40000. There is a
typo in the existing reg values. Correct the same. It doesnt have any functional
impact as the HIGH_ADDRESS is not being used anywhere in the driver sources.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>

Acked-for-series: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>