device_tree: Correct the usage of set_dt_param for include_dts and repo options
Correct the usage of set_dt_param for include_dts and repo options. Present
implementation doesnt allow the usage of multiple subsequent usage of set_dt_param
for include_dts and repo. As the default value of env(REPO) and env(include_dts)
are set everytime, the user set values get overwritten in the subsequent
set_dt_param calls. Fix the issue by introducing a new proc which will be called
during generate_sdt and get_dt_param and will set the default value only when
the value is not already set.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>
Acked-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>