axi_traffic_gen: Generate atg-mode properly when it is AXI4-Stream
The properties atg-mode, atg-mode-l2 and axis-mode are getting generated
incorrectly when atg-mode is AXI4-Stream and it is due to wrongly placed
return condition in the tcl. Hence, move the if condition and return to
appropriate place in the tcl so that these properties can be generated
properly in all scenarios.

Signed-off-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>
Acked-by: Onkar Harsh <onkar.harsh@amd.com>