common_proc: Update logic to check if the address is defined in the static dtsi
PL IPs in Versal SOC designs can be assigned the same address as that of a PS IP.
In such scenarios these PL IPs are only accessible by the Soft Microblaze.
APU/RPU can still access the respective PS IP from the same address. Therefore,
update the existing logic to represent the same in the system device tree.

Add a check if the IP under test is a PL IP and move ahead with further processing
for such IPs even if the base address of IP is found under the ps_mapping dict.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>
Acked-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>