common_proc: Update the gen_reg_property proc to generate PS specific mapping first
In case of PS+PL designs, Force the PS mapped memory range to appear first in
list. It has been observed in a design that an IP is mapped to both PS and PL
processor and is having more than 32bit address when mapped to PS, but the same
IP's address (as obtained from HSI) when mapped to PL processor is truncated to
lower 32 bits and is not intended to be used with PL. Baremetal always takes the
first reg instance from the node unless specified otherwise and having PL mapped
address first leads to wrong base address for such IPs in the BSP which is
targeted for PS processor.

Also maintain a global variable "apu_proc_ip" under set_hw_family proc to avoid
using HSI command with filter everytime the PS processor IP name is needed.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>

Acked-for-series: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>