common: Update the definition of get_baseaddr and get_highaddr
get_baseaddr and high_addr procs always return the BASE_VALUE and HIGH_VALUE of
0th instance of get_mem_ranges output for the IP under test irrespective of the
processor mapping for that 0th instance. This is leading to wrong mapping in
processor address maps. Provide an option to get the processor handle as an input
and return only the memory instances mapped to the particular processor.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>

Acked-for-series: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>