axi_bram: Add condition to add BRAM memory in the Zynq APU address map
BRAM acts as a pure memory IP when the ECC Flag is not set in the IP. Address-maps
of all such memory IPs are set in their respective TCLs. While setting the memory
map, Zynq APU related mapping was missing. Add the condition to map the same.

In addition, optimize multiple if else conditions with the recently added global
APU variable.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>
Acked-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>