ipipsu: Update logic to generate xlnx, ipi-id values in IPI nodes similar to ATF
In versal and versal-net, the xlnx,ipi-id is based on bit position and
it is same as ATF because the bit positions of IPI channels are
contiguous.
In zynqmp the bit positions of IPI channels are not contiguous as there
are few reserved bits in between the channels which causes to generate
xlnx,ipi-id values different from that of ATF.

Update logic for zynqmp to generate xlnx,ipi-id values in IPI nodes
similar to that of ATF by using dictionary baseaddress-IPI ID mapping.

Signed-off-by: Harini T <harini.t@amd.com>
Acked-by: Srinivas Goud <srinivas.goud@amd.com>