common: Use the same static node for psv_cpm versal as that of linux
Base address of psv_cpm IP (CPM4 and CPM5) being used in Linux DT is not same as
the one provided by HSI. There is a different mechanism to reach to the same base
address as that of the Linux DT. Add support for the same.

Also, add node for CPM5 in versal static dtsi file.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>
Signed-off-by: Bhavana Jupalli <Bhavana.Jupalli@amd.com>

Acked-for-series: Bharat Kumar Gogada <bharat.kumar.gogada@amd.com>