mig7_series: Check the 32-bit cell size boundary address to avoid overflow
4 GB (0xFFFFFFFF) is a boundary condition for a 32-bit cell size. Size is
calculated using (high_addr - low_addr + 1) in SDT. When the complete 4 GB DDR
is used for a 32 bit processor, the size becomes 0x100000000 using the above
calculation and result into an overflow for a 32-bit cell size in SDT. Therefore,
avoid adding 1 in size for this boundary condition.

In addition, Consolidate some if conditions to make the code look a bit cleaner.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>

Acked-for-series: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>