common_proc: check the presence of intc IP before using its reference
For preset_wrapper designs, the intc IP is not mapped to the microblaze processor
but it appears in get_cells -hier ouput. Moreover, the intc may not have the base
address in such cases which will lead to missing reference for intc in the SDT.
Therefore, check the presence of intc before using its reference in cpu node as
interrupt handler.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>

Acked-for-series: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>