device_tree: Fix the missing interrupt property in child nodes
With the current implementation only two of the mandatory three
interrupt properties for the child node of DMA are only defined.
This is because peripheral interrupt controller generate multiple
interrupt instances for the IPs which have interrupt properties in
the child node. Fix by setting a variable that holds the IP name
of the peripheral interrupt controller for every interrupt.

Signed-off-by: Abin Joseph <abin.joseph@amd.com>
Acked-by: Onkar Harsh <onkar.harsh@amd.com>