common: Correct the frequencies of clk_cpu nodes when multiple MBs are present in design
There is a node cpu_clk which is legacy in nature and is generated for the
soft microblaze of the design. In designs where we have multiple MBs available,
the frequency representation within this node goes wrong. With existing
implementation, only one clk_cpu node exists irrespective of the number of MBs
in the design and adding multiple frequencies to the property of the same node
results in the syntax error. Extend the logic to have correct unit name and
clk_cpu node name for each of the microblazes available in the design.

The real purpose of clk_cpu is not clear and it is a legacy code taken up from
DTG. This needs to be revisited while adding Linux support for Soft Microblaze
in SDT flow.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>

Acked-for-series: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>