common: Use MASTER_INTERFACE of each memory object to determine the base address
For versal designs, the 0th index of the get_mem_ranges of some of the IPs may
not be the correct base address to be put in system device tree. For example,
pmc IPs like psv_pmc_qspi, psv_dma, psv_coresight can have the master interface
as CPM_PCIE_NOC_0 or FPD_AXI_NOC_1. These addresses are not part of processor
views and need to be ignored while setting the base address for the IP instances.

Therefore, add logic to check if the master interface of the memory object of each
IP instance starts with any of the available processor names. This case is valid
only for PS IPs as PL IPs master interfaces dont have processor entries in them.
Moreover, to minimize the effect of this change, add this only for versal designs.

In addition, optimize the get_baseaddr and high_baseaddr usage by introducing a
global dictionary for each of them.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>
Acked-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>