common: Add a check to detect overflow of handle_counts while deriving base addresses
There are few IPs (namely M_AXIL, ps_pl_axil) which are not coming as part of
hsi get_cells -hier but they are appearing in hsi get_mem_ranges -of_objects <apu_0>.
This issue is solved earlier for CCAT kind of INTERFACE IPs but that time the
IP had a valid base address. This time, these IPs are coming under memory objects
for apu_0 but none of the instances have MASTER_INTERFACE as APU. (MASTER INTERFACE
is FPD_AXI_NOC_0 and CPM_PCIE_NOC_1 for all the instances). Thus, base address
processing resulted into overflow of index values for the list under this
particular condition.

Add a check for this index overflow and return invalid base address. Validate
the base address for this condition before creating nodes for such IPs.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>
Acked-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>