device_tree: Correct the configuration read format for VP1902 device in SDT flow.
For VP1902 device, When XilSEM is not enabled in vivado,
the XilSEM macros (XSEM_CFRSCAN_EN, XSEM_NPISCAN_EN) are
populated in xparameters.h file.
This is due to change in configuration format for VP1902
device. We are getting CRAM EN config string as list
"0 0 0 0" instead of a single integer / string.
Correct configuration reading format for VP1902 in SDT
flow.

Signed-off-by: GayathriM <gayathri.m@amd.com>
Acked-by: Onkar Harsh <onkar.harsh@amd.com>