ddrpsv: Check axi noc IP_NAME instead of the platform to get the DDR address blocks
The address block name depends on the IP and not on the platform. For axi_noc,
there is a set of block names and for axi_noc2, there is another set of block
names. Presently, the condition relies on platform check whether it is versal or
versal net. Update the condition to check the IP_NAME instead.

Additionally, update the proc name accordingly. And fix the conditions that
compares the base addr and high addre size for reg property calculation. string
comparison is leading to wrong comparison, convert the same to the hex comparison.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>

Acked-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>