common_proc: Loop over interrupt pins to decide if it is connected to an INTR_CNTRL
For TMR designs, there can be interrupts going to comparator IP blocks of TMR
which are not really a source of INTERRUPTS for Software but are shown as INTERRUPT
pins in the design. Loop over all the connected pins and break when a controller is
found. Before encountering this design, connectip was always returning single
instance for all the known designs.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>
Acked-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>