common: unset duplicate mapping from device tree
For a variant of versal 2VE and 2VM devices, the gpu base address is allocated to a different IP and the gpu is absent in the device. Due to this the node reference for this new IP is getting set as that of gpu in the pcw.dtsi, which is causing run time probe errors for gpu. Therefore, unset that gpu address from ps_mapping list in the parts_specific file.

Signed-off-by: Dhriti Sree Rompicharla <DhritiSree.Rompicharla@amd.com>
Acked-by: Onkar Harsh <onkar.harsh@amd.com>